requirements.txt
Django==3.0.1
gunicorn==20.0.4
djangorestframework==3.11.0
markdown==3.1.1
django-filter==2.2.0